Rename "bootstrap" to "migrate" mode
Fixes #18
